+++ src/disc_fdi.h
@@ -10,7 +10,7 @@ void fdi_writesector(int drive, int sector, int track, int side, int density, in
 void fdi_readaddress(int drive, int sector, int side, int density);
 void fdi_format(int drive, int sector, int side, int density, uint8_t fill);
 int fdi_hole(int drive);
-int fdi_byteperiod(int drive);
+double fdi_byteperiod(int drive);
 void fdi_stop();
 void fdi_poll();
 int fdi_realtrack(int track, int drive);
